Add unscoped option to through association
The main reason for the join table is for when two different schools "create" a category with the same name; we want to avoid having duplicated categories across schools, because this puts the onus on the deduplication to happen everywhere it's needed at runtime, e.g. reporting, district dashboards. If a school adds a category that already exists at another school, we simply create a new join to the same category record